fix(gateway): claim eviction victims and kill only after replacement spawn
