+++ src/gateway/terminal/session-manager.test.ts
@@ -1,11 +1,12 @@
 import { expectDefined } from "@openclaw/normalization-core";
 import { describe, expect, it, vi } from "vitest";
-import type { spawnTerminalPty } from "../../process/terminal-pty.js";
 import type { TerminalBackend } from "./backend.js";
 import { TerminalSessionManager } from "./session-manager.js";
-
-type TerminalOpenRequest = Parameters<TerminalSessionManager["open"]>[0];
-type TerminalPtyHandle = Awaited<ReturnType<typeof spawnTerminalPty>>;
+import {
+  baseOpenRequest as baseRequest,
+  type FakeTerminalPty,
+  makeFakePty,
+} from "./session-manager.test-helpers.js";
 const TERMINAL_EVENT_DATA = "terminal.data";
 const TERMINAL_EVENT_EXIT = "terminal.exit";
 
@@ -17,76 +18,9 @@ function deferred<T>() {
   return { promise, resolve };
 }
 
-/** A controllable fake PTY that records writes and lets tests drive data/exit. */
-function makeFakePty() {
-  let dataListener: ((chunk: string) => void) | undefined;
-  let exitListener: ((event: { exitCode: number; signal?: number }) => void) | undefined;
-  const handle: TerminalPtyHandle & {
-    writes: string[];
-    resizes: Array<[number, number]>;
-    killed: boolean;
-    paused: boolean;
-    pauseCalls: number;
-    resumeCalls: number;
-    deliveredChunks: number;
-    emitData: (chunk: string) => void;
-    emitExit: (code: number, signal?: number) => void;
-  } = {
-    pid: 4242,
-    writes: [],
-    resizes: [],
-    killed: false,
-    paused: false,
-    pauseCalls: 0,
-    resumeCalls: 0,
-    deliveredChunks: 0,
-    write: (data) => handle.writes.push(data),
-    resize: (cols, rows) => handle.resizes.push([cols, rows]),
-    pause: () => {
-      handle.paused = true;
-      handle.pauseCalls += 1;
-    },
-    resume: () => {
-      handle.paused = false;
-      handle.resumeCalls += 1;
-    },
-    onData: (listener) => {
-      dataListener = listener;
-    },
-    onExit: (listener) => {
-      exitListener = listener;
-    },
-    kill: () => {
-      handle.killed = true;
-    },
-    emitData: (chunk) => {
-      if (!handle.paused) {
-        handle.deliveredChunks += 1;
-        dataListener?.(chunk);
-      }
-    },
-    emitExit: (code, signal) => exitListener?.({ exitCode: code, signal }),
-  };
-  return handle;
-}
-
-function baseRequest(overrides?: Partial<TerminalOpenRequest>): TerminalOpenRequest {
-  return {
-    owner: { kind: "conn", connId: "conn-1" },
-    agentId: "main",
-    cwd: "/work",
-    shell: "/bin/zsh",
-    args: ["-l"],
-    cols: 80,
-    rows: 24,
-    env: { TERM: "xterm-256color" },
-    ...overrides,
-  };
-}
-
 describe("TerminalSessionManager", () => {
   it("kills a backend that finishes after its open request is cancelled", async () => {
-    const spawned = deferred<TerminalPtyHandle>();
+    const spawned = deferred<FakeTerminalPty>();
     const controller = new AbortController();
     const first = makeFakePty();
     const second = makeFakePty();
@@ -118,8 +52,8 @@ describe("TerminalSessionManager", () => {
   });
 
   it("bounds cancelled backend operations until they settle", async () => {
-    const firstSpawn = deferred<TerminalPtyHandle>();
-    const secondSpawn = deferred<TerminalPtyHandle>();
+    const firstSpawn = deferred<FakeTerminalPty>();
+    const secondSpawn = deferred<FakeTerminalPty>();
     const firstController = new AbortController();
     const secondController = new AbortController();
     let spawnCount = 0;
@@ -692,67 +626,6 @@ describe("TerminalSessionManager agent ownership", () => {
     } finally {
       vi.useRealTimers();
     }
-  });
-
-  it("evicts the longest-idle viewer-free agent session under pool pressure", async () => {
-    vi.useFakeTimers();
-    try {
-      const ptys: ReturnType<typeof makeFakePty>[] = [];
-      const manager = new TerminalSessionManager({
-        emit: vi.fn(),
-        spawn: async () => {
-          const pty = makeFakePty();
-          ptys.push(pty);
-          return pty;
-        },
-        maxSessions: 2,
-      });
-      const first = await manager.open(baseRequest({ owner: agentOwner }));
-      await vi.advanceTimersByTimeAsync(5_000);
-      const second = await manager.open(baseRequest({ owner: agentOwner }));
-      if (!first.ok || !second.ok) {
-        throw new Error("expected agent opens");
-      }
-      // Freshen the second session so the first is the idle-eviction candidate.
-      await vi.advanceTimersByTimeAsync(5_000);
-      expect(manager.writeAgent("agent:main:main", second.sessionId, "keepalive\r")).toBe(true);
-
-      const third = await manager.open(baseRequest({ owner: agentOwner }));
-      expect(third.ok).toBe(true);
-      expect(manager.size).toBe(2);
-      expect(expectDefined(ptys[0], "ptys[0] test invariant").killed).toBe(true);
-      expect(expectDefined(ptys[1], "ptys[1] test invariant").killed).toBe(false);
-      expect(manager.snapshotAgent("agent:main:main", first.sessionId)).toBeUndefined();
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
-  it("never evicts viewer-attached or connection-owned sessions under pressure", async () => {
-    const ptys: ReturnType<typeof makeFakePty>[] = [];
-    const manager = new TerminalSessionManager({
-      emit: vi.fn(),
-      spawn: async () => {
-        const pty = makeFakePty();
-        ptys.push(pty);
-        return pty;
-      },
-      maxSessions: 2,
-    });
-    const connOwned = await manager.open(baseRequest());
-    const viewed = await manager.open(baseRequest({ owner: agentOwner }));
-    if (!connOwned.ok || !viewed.ok) {
-      throw new Error("expected opens");
-    }
-    expect(manager.attach("viewer-1", viewed.sessionId)?.sessionId).toBe(viewed.sessionId);
-
-    const denied = await manager.open(baseRequest({ owner: agentOwner }));
-    expect(denied.ok).toBe(false);
-    if (!denied.ok) {
-      expect(denied.code).toBe("limit");
-    }
-    expect(manager.size).toBe(2);
-    expect(ptys.some((pty) => pty.killed)).toBe(false);
   });
 
   it("lets a co-attached viewer upload into an agent-owned session", async () => {
